+++ NEWS.md
@@ -3,6 +3,7 @@ Version 2.0.11
 
 * New `ranef` S3 generic and `ranef.glmmVA` method: extracts random-effect estimates from a `glmmVA` object.
 * New multiple response types support for `glmmVA`
+* Distance matrix for spatial models is now calculated once in R, in contast to repeatedly in C++
 
 ## Bugfixes